Added basic flow test to Message Inbox

+++ packages/hoprd/crates/hoprd-inbox/src/inbox.rs
@@ -32,12 +32,16 @@ pub trait InboxBackend<T: Copy + Default, M> {
     async fn purge(&mut self, older_than_ts: Duration);
 }
 
+/// Holds basic configuration parameters of the `MessageInbox`.
 #[derive(Debug, Clone, Eq, PartialEq)]
 #[cfg_attr(feature = "wasm", wasm_bindgen::prelude::wasm_bindgen(getter_with_clone))]
 pub struct MessageInboxConfiguration {
     /// Maximum capacity per-each application tag.
+    /// In the current implementation, the capacity must be a power of two.
     pub capacity: u32,
+    /// Maximum age of a message held in the inbox until it is purged.
     pub max_age_sec: u64,
+    /// List of tags that are excluded on `push`.
     pub excluded_tags: Vec<Tag>,
 }
 
@@ -53,6 +57,7 @@ impl Default for MessageInboxConfiguration {
 
 /// Represents a thread-safe message inbox of messages of type `M`
 /// This type is thin frontend over `InboxBackend` using 16-bit unsigned integer tags.
+/// Each operation also performs `purge` of the backend.
 pub struct MessageInbox<B>
 where
     B: InboxBackend<Tag, ApplicationData>,
@@ -66,6 +71,9 @@ impl<B> MessageInbox<B>
 where
     B: InboxBackend<Tag, ApplicationData>,
 {
+    /// Creates new instance given the configuration.
+    /// Uses `std::time::SystemTime` as timestamping function, therefore
+    /// this function cannot be used in WASM runtimes.
     #[cfg(any(not(feature = "wasm"), test))]
     pub fn new(cfg: MessageInboxConfiguration) -> Self {
         Self::new_with_time(cfg, || {
@@ -75,6 +83,7 @@ where
         })
     }
 
+    /// Creates new instance given the configuration and the timestamping function.
     pub fn new_with_time(cfg: MessageInboxConfiguration, time: TimestampFn) -> Self {
         Self {
             backend: Mutex::new(B::new_with_capacity(cfg.capacity as usize, time)),
@@ -83,35 +92,63 @@ where
         }
     }
 
-    pub async fn push(&self, payload: ApplicationData) -> bool {
-        if payload.application_tag.is_none()
-            || payload
-                .application_tag
-                .iter()
-                .all(|e| e.eq(&payload.application_tag.unwrap()))
-        {
-            let mut db = self.backend.lock().await;
-            db.push(payload.application_tag, payload).await;
-            db.purge((self.time)() - Duration::from_secs(self.cfg.max_age_sec))
-                .await;
-            true
-        } else {
-            false
+    fn is_excluded_tag(&self, tag: &Option<Tag>) -> bool {
+        match tag {
+            None => false,
+            Some(t) => self.cfg.excluded_tags.iter().any(|e| e.eq(t))
         }
     }
 
-    pub async fn size(&self, tag: Option<Tag>) -> usize {
-        self.backend.lock().await.count(tag).await
+    /// Push message into the inbox. Returns `true` if the message has been enqueued, `false` if it
+    /// has been excluded based on the configured exluded tags.
+    pub async fn push(&self, payload: ApplicationData) -> bool {
+        if self.is_excluded_tag(&payload.application_tag) {
+            return false
+        }
+
+        // Push only if there is no tag, or if the tag is not excluded
+        let mut db = self.backend.lock().await;
+        db.push(payload.application_tag, payload).await;
+        db.purge((self.time)() - Duration::from_secs(self.cfg.max_age_sec))
+            .await;
+
+        true
     }
 
+    /// Number of messages in the inbox with the given `tag`, or the total number of all messages
+    /// if no `tag` is given.
+    pub async fn size(&self, tag: Option<Tag>) -> usize {
+        if self.is_excluded_tag(&tag) {
+            return 0
+        }
+
+        let mut db = self.backend.lock().await;
+        db.purge((self.time)() - Duration::from_secs(self.cfg.max_age_sec))
+            .await;
+        db.count(tag).await
+    }
+
+    /// Pop the oldest message with the given tag, or the oldest message regardless the tag
+    /// if it is not given. Returns `None` if there's no message with such `tag` (if given) in the inbox
+    /// or if the whole inbox is empty (if no `tag` is given).
     pub async fn pop(&self, tag: Option<Tag>) -> Option<ApplicationData> {
+        if self.is_excluded_tag(&tag) {
+            return None
+        }
+
         let mut db = self.backend.lock().await;
         db.purge((self.time)() - Duration::from_secs(self.cfg.max_age_sec))
             .await;
         db.pop(tag).await
     }
 
+    /// Pops all the messages with the given `tag` (ordered oldest to latest) or
+    /// all the messages from the entire inbox (ordered oldest to latest) if no `tag` is given.
     pub async fn pop_all(&self, tag: Option<Tag>) -> Vec<ApplicationData> {
+        if self.is_excluded_tag(&tag) {
+            return Vec::new()
+        }
+
         let mut db = self.backend.lock().await;
         db.purge((self.time)() - Duration::from_secs(self.cfg.max_age_sec))
             .await;
@@ -121,9 +158,43 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::time::Duration;
+    use core_packet::interaction::{ApplicationData, Tag};
+    use crate::inbox::{MessageInbox, MessageInboxConfiguration};
+    use crate::ring::RingBufferInboxBackend;
 
-    #[test]
-    fn test_basic_flow() {}
+    #[async_std::test]
+    async fn test_basic_flow() {
+        let cfg = MessageInboxConfiguration {
+            capacity: 4,
+            excluded_tags: vec![2],
+            max_age_sec: 2
+        };
+
+        let mi = MessageInbox::<RingBufferInboxBackend<Tag, ApplicationData>>::new(cfg);
+
+        assert!(mi.push(ApplicationData { application_tag: None, plain_text: (*b"test msg 0").into() }).await);
+        assert!(mi.push(ApplicationData { application_tag: Some(1), plain_text: (*b"test msg 1").into() }).await);
+        assert!(mi.push(ApplicationData { application_tag: Some(1), plain_text: (*b"test msg 2").into() }).await);
+        assert!(! mi.push(ApplicationData { application_tag: Some(2), plain_text: (*b"test msg").into() }).await);
+        assert_eq!(3, mi.size(None).await);
+        assert_eq!(2, mi.size(Some(1)).await);
+        assert_eq!(0, mi.size(Some(2)).await);
+
+        let ad = mi.pop(None).await.unwrap();
+        assert_eq!(b"test msg 0", ad.plain_text.as_ref());
+
+        let ad = mi.pop(Some(1)).await.unwrap();
+        assert_eq!(b"test msg 1", ad.plain_text.as_ref());
+        assert_eq!(1, mi.size(Some(1)).await);
+
+        assert_eq!(1, mi.size(None).await);
+
+        async_std::task::sleep(Duration::from_millis(2500)).await;
+
+        assert_eq!(0, mi.size(None).await);
+
+    }
 }
 
 #[cfg(feature = "wasm")]
